List Layers for a Given Organization

This function retrieves and lists the available layers for a specified organization.

Usage

argentum_list_layers(organization)

Arguments

organization

A character string specifying the name of the organization

Value

A data frame containing layer information with columns 'Name' and 'Title'

Examples

# \donttest{
# This example uses a mock organization name.
# In a real scenario, use an actual organization name from argentum_list_organizations()
tryCatch({
  orgs <- argentum_list_organizations()
  if(nrow(orgs) > 0) {
    layers <- argentum_list_layers(orgs$Name[1])
    print(layers)
  } else {
    message("No organizations found.")
  }
}, error = function(e) {
  message("Error occurred: ", e$message)
})
#> Attempting to connect to WFS URL: https://geotematico01.conae.gov.ar/geoserver/wfs?service=WFS&version=1.1.0&request=GetCapabilities
#> Requesting capabilities from: https://geotematico01.conae.gov.ar/geoserver/wfs?service=WFS&version=1.1.0&request=GetCapabilities
#>                                             Name
#> 1         Localidades:Centro_huellas_localidades
#> 2           Localidades:Cuyo_huellas_localidades
#> 3            Localidades:NEA_huellas_localidades
#> 4            Localidades:NOA_huellas_localidades
#> 5 Localidades:PatagoniaNorte_huellas_localidades
#> 6   Localidades:PatagoniaSur_huellas_localidades
#> 7               Localidades:imagenes_disponibles
#>                                      Title
#> 1          Centro - Huellas de localidades
#> 2            Cuyo - Huellas de localidades
#> 3             NEA - Huellas de localidades
#> 4             NOA - Huellas de localidades
#> 5 Patagonia Norte - Huellas de localidades
#> 6   Patagonia Sur - Huellas de localidades
#> 7                     imagenes_disponibles
# }