Select Organization

Select Organization

Usage

argentum_select_organization(search = NULL, interactive_select = interactive())

Arguments

search

An optional search term to filter organizations

interactive_select

Logical; whether to run in interactive mode. Defaults to interactive()

Value

A data frame containing the selected organization's information, or NULL if no selection is made

Examples

# List available organizations
orgs <- argentum_list_organizations()
print(orgs)
#> # A tibble: 130 × 3
#>    Name                                                          WMS_URL WFS_URL
#>    <chr>                                                         <chr>   <chr>  
#>  1 "Jefatura de Gabinete de Ministros - Comisión Nacional de Ac… https:… https:…
#>  2 "Jefatura de Gabinete de Ministros - Comisión Nacional de Ac… https:… https:…
#>  3 "Jefatura de Gabinete de Ministros - Comisión Nacional de Ac… https:… https:…
#>  4 "Jefatura de Gabinete de Ministros - Subsecretaría de Teleco… https:… https:…
#>  5 "Ministerio de Defensa - Instituto Geográfico Nacional (IGN)" http:/… https:…
#>  6 "Ministerio de Defensa - Ministerio de Defensa"               https:… https:…
#>  7 "Ministerio de Defensa - Servicio de Hhidrografía Naval (SHN… http:/… http:/…
#>  8 "Ministerio de Economía - Instituto Nacional de Estadística … https:… https:…
#>  9 "Ministerio de Economía - Instituto Nacional de Tecnología A… http:/… http:/…
#> 10 "Ministerio de Economía - Secretaría de Bioeconomía (Agricul… https:… https:…
#> # ℹ 120 more rows

# Select organization non-interactively
selected_org <- argentum_select_organization(interactive_select = FALSE)
#> Available organizations:
#>   [1] "Jefatura de Gabinete de Ministros - Comisión Nacional de Actividades Espaciales (CONAE) / Imágenes satelitales- Mosaico Alta Resolución   "   
#>   [2] "Jefatura de Gabinete de Ministros - Comisión Nacional de Actividades Espaciales (CONAE) / Riesgo epidemiológico - Riesgo ambiental    "       
#>   [3] "Jefatura de Gabinete de Ministros - Comisión Nacional de Actividades Espaciales (CONAE) / Suelos -Focos de calor"                             
#>   [4] "Jefatura de Gabinete de Ministros - Subsecretaría de Telecomunicaciones y Conectividad (IDECOM)"                                              
#>   [5] "Ministerio de Defensa - Instituto Geográfico Nacional (IGN)"                                                                                  
#>   [6] "Ministerio de Defensa - Ministerio de Defensa"                                                                                                
#>   [7] "Ministerio de Defensa - Servicio de Hhidrografía Naval (SHN)"                                                                                 
#>   [8] "Ministerio de Economía - Instituto Nacional de Estadística y Censos (INDEC)"                                                                  
#>   [9] "Ministerio de Economía - Instituto Nacional de Tecnología Agropecuaria (INTA)"                                                                
#>  [10] "Ministerio de Economía - Secretaría de Bioeconomía (Agricultura)"                                                                             
#>  [11] "Ministerio de Economía - Secretaría de Energía (ex Subsecretaría de Escenarios y Evaluación de Proyectos del Ministerio de Energía y Minería)"
#>  [12] "Ministerio de Economía - Secretaría de Obras Públicas"                                                                                        
#>  [13] "Ministerio de Economía - Secretaría de Transporte"                                                                                            
#>  [14] "Ministerio de Economía - Secretaría de Industria y Comercio"                                                                                  
#>  [15] "Ministerio de Economía - Servicio Nacional de Sanidad y Calidad Agroalimentaria (SENASA)"                                                     
#>  [16] "Ministerio de Economía - Sigvial - Dirección Nacional de Vialidad"                                                                            
#>  [17] "Ministerio de Economía - Sistema de Información Geográfica del Servicio Geológico Minero Argentino (SEGEMAR)"                                 
#>  [18] "Ministerio de Economía, Ministerio de Educación, Ministerio de Defensa - BAHRA"                                                               
#>  [19] "Ministerio de Educación - Programa Nacional Mapa Educativo"                                                                                   
#>  [20] "Ministerio de Infraestructura - Instituto Nacional del Agua (INA)"                                                                            
#>  [21] "Ministerio del Interior - Dirección Nacional de Población (RENAPER)"                                                                          
#>  [22] "Ministerio del Interior - Subsecretaría de Ambiente "                                                                                         
#>  [23] "Buenos Aires - Infraestructura de Datos Espaciales de Buenos Aires (IDEBA)"                                                                   
#>  [24] "Buenos Aires - Agencia de Recaudación de la Provincia de Buenos Aires (ARBA)"                                                                 
#>  [25] "Buenos Aires - Mapa escolar"                                                                                                                  
#>  [26] "Buenos Aires - Provincia de Buenos Aires-Infraestructura de Datos Espaciales"                                                                 
#>  [27] "Buenos Aires - URBASIG"                                                                                                                       
#>  [28] "Buenos Aires - Dirección Provincial de Estadística"                                                                                           
#>  [29] "Buenos Aires - Organismo Provincial de Integración Social y Urbana"                                                                           
#>  [30] "Buenos Aires - Sistema de Análisis Territorial Ambiental (SATA)"                                                                              
#>  [31] "Buenos Aires - Ministerio de Justicia y Derechos Humanos"                                                                                     
#>  [32] "Buenos Aires - Ministerio de las Mujeres, Políticas de Género y Diversidad Sexual"                                                            
#>  [33] "Buenos Aires - Registro Provincial de Organizaciones de la Comunidad (REPOC)"                                                                 
#>  [34] "Buenos Aires - Ministerio de Producción, Ciencia e Innovación Tecnológica"                                                                    
#>  [35] "Catamarca - Provincia de Catamarca (IDECAT)"                                                                                                  
#>  [36] "Catamarca - Ministerio de Salud"                                                                                                              
#>  [37] "Catamarca - Ministerio de Educación"                                                                                                          
#>  [38] "Catamarca - Ministerio de Agua, Energía y Medio Ambiente"                                                                                     
#>  [39] "Catamarca - Estadísticas"                                                                                                                     
#>  [40] "Catamarca - Mapa Productivo"                                                                                                                  
#>  [41] "Catamarca - Ministerio de Seguridad"                                                                                                          
#>  [42] "Chaco - Provincia del Chaco (IDECHACO)"                                                                                                       
#>  [43] "Chubut - IDE Estadísticas y Censos"                                                                                                           
#>  [44] "Chubut - Provincia del Chubut (IDECHUBUT)"                                                                                                    
#>  [45] "Chubut - Sistema de Información Territorial"                                                                                                  
#>  [46] "Ciudad de Buenos Aires - Dirección General de Estadísticas y Censos"                                                                          
#>  [47] "Corrientes - Provincia de Corrientes (IDECorr)"                                                                                               
#>  [48] "Córdoba - Provincia de Córdoba (IDECOR)"                                                                                                      
#>  [49] "Entre Ríos - Provincia de Entre Ríos (IDEER)"                                                                                                 
#>  [50] "Jujuy - Ministerio de infraestructura, servicios Públicos, Tierra y Vivienda"                                                                 
#>  [51] "La Pampa - Administración Provincial de Energia  - APE"                                                                                       
#>  [52] "La Pampa - Dirección General de Catastro - Vuelos VANT"                                                                                       
#>  [53] "La Pampa - Dirección Provincial de Vialidad"                                                                                                  
#>  [54] "La Pampa - Dirección de Defensa Civil"                                                                                                        
#>  [55] "La Pampa - Ministerio de Hacienda y Finanzas - Dirección General de Catastro"                                                                 
#>  [56] "La Pampa - Ministerio de Salud"                                                                                                               
#>  [57] "La Pampa - Provincia de La Pampa (IDELP)"                                                                                                     
#>  [58] "La Pampa - Secretaría de Cultura"                                                                                                             
#>  [59] "La Pampa - Secretaría de Turismo"                                                                                                             
#>  [60] "La Pampa - Subsecretaría de Ambiente"                                                                                                         
#>  [61] "Mendoza - Dirección General de Catastro"                                                                                                      
#>  [62] "Mendoza - Secretaría de Ambiente y Ordenamiento Territorial (SIAT)"                                                                           
#>  [63] "Misiones - Electricidad de Misiones Sociedad Anónima (EMSA)"                                                                                  
#>  [64] "Misiones - Fotografías Aéreas - 2012"                                                                                                         
#>  [65] "Misiones - Instituto Provincial de Desarrollo Habitacional (IPRODHA)"                                                                         
#>  [66] "Misiones - Instituto Provincial de Estadísticas y Censo (IPEC)"                                                                               
#>  [67] "Misiones - Instituto Provincial de Lotería y Casinos (IPLyC)"                                                                                 
#>  [68] "Misiones - Min. de Ecología y Recursos Naturales Renovables"                                                                                  
#>  [69] "Misiones - Ministerio de Cultura, Educación Ciencia y Tecnología"                                                                             
#>  [70] "Misiones - Ministerio de Desarrollo Social, La mujer y la Juventud"                                                                           
#>  [71] "Misiones - Ministerio de Gobierno"                                                                                                            
#>  [72] "Misiones - Ministerio de Salud Pública"                                                                                                       
#>  [73] "Misiones - Ministerio de Turismo"                                                                                                             
#>  [74] "Misiones - Ministerio del Agro y la Producción"                                                                                               
#>  [75] "Misiones - Provincia de Misiones"                                                                                                             
#>  [76] "Misiones - Provincia de Misiones - Instituto Provincial de Desarrollo Habitacional"                                                           
#>  [77] "Neuquén - Autoridad Interjurisdiccional de Cuencas (AIC)"                                                                                     
#>  [78] "Neuquén - COPADE"                                                                                                                             
#>  [79] "Neuquén - Catastro"                                                                                                                           
#>  [80] "Neuquén - Defensa Civil"                                                                                                                      
#>  [81] "Neuquén - Educación"                                                                                                                          
#>  [82] "Neuquén - Estadistica"                                                                                                                        
#>  [83] "Neuquén - Hidrocarburos"                                                                                                                      
#>  [84] "Neuquén - Hidrocarburos"                                                                                                                      
#>  [85] "Neuquén - IPVU-ADUS (Barrios IPVU-ADUS)"                                                                                                      
#>  [86] "Neuquén - IPVU-ADUS (Parcelas provisorias)"                                                                                                   
#>  [87] "Neuquén - IPVU-ADUS (Viviendas)"                                                                                                              
#>  [88] "Neuquén - Ministerio de Producción y Ambiente"                                                                                                
#>  [89] "Neuquén - Oficina Provincial de Tecnologias de la Informacion y la Comunicación (OPTIC)"                                                      
#>  [90] "Neuquén - Policía"                                                                                                                            
#>  [91] "Neuquén - Recursos Hídricos"                                                                                                                  
#>  [92] "Neuquén - Vialidad"                                                                                                                           
#>  [93] "Río Negro - Infraestructura de Datos Espaciales de Río Negro (IDERN)"                                                                         
#>  [94] "Salta - Infraestructura de Datos Espaciales de Salta (IDESA)"                                                                                 
#>  [95] "Santa Cruz - Infraestructura de Datos Espaciales de Santa Cruz (SITU)"                                                                        
#>  [96] "Tucumán - Secretaría de Estado de Producción- Indicadores sociodemográficos"                                                                  
#>  [97] "Tucumán - Dirección de Estadística de la Provincia"                                                                                           
#>  [98] "Tucumán - Dirección General de Catastro"                                                                                                      
#>  [99] "Tucumán - Centro de Estudios del Territorio y Hábitat Popular - FAU - UNT"                                                                    
#> [100] "Tucumán - Secretaría de Estado de Producción- Turismo"                                                                                        
#> [101] "Tucumán - Secretaría de Estado de Producción- Información productiva"                                                                         
#> [102] "Tucumán - Secretaría de Estado de Producción- Ambientes naturales"                                                                            
#> [103] "Tucumán - Secretaría de Estado de Producción- Infraestructura"                                                                                
#> [104] "Tucumán - Secretaría de Estado de Producción- Areas productivas"                                                                              
#> [105] "Buenos Aires - Municipalidad de Balcarce"                                                                                                     
#> [106] "Buenos Aires - Municipalidad de Benito Juarez"                                                                                                
#> [107] "Buenos Aires - Municipalidad de Luján"                                                                                                        
#> [108] "Buenos Aires - Municipalidad de Moreno"                                                                                                       
#> [109] "Buenos Aires - Municipalidad de Olavarría"                                                                                                    
#> [110] "Buenos Aires - Municipalidad de Pergamino "                                                                                                   
#> [111] "Catamarca - Municipalidad de Fray Mamerto Esquiú "                                                                                            
#> [112] "Catamarca - Municipalidad de San Fernando del Valle de Catamarca"                                                                             
#> [113] "Catamarca - Municipalidad de Tinogasta"                                                                                                       
#> [114] "Córdoba - Municipalidad de Villa María"                                                                                                       
#> [115] "Corrientes - Municipalidad de Corrientes"                                                                                                     
#> [116] "Entre Ríos - Municipalidad de San Benito"                                                                                                     
#> [117] "La Pampa - Municipalidad de Santa Rosa"                                                                                                       
#> [118] "Misiones - Municipalidad de Posadas"                                                                                                          
#> [119] "Salta - Municipalidad de Salta"                                                                                                               
#> [120] "Universidad de Buenos Aires - Facultad de Ingeniería de la Universidad de Buenos Aires (FIUBA)- Departamento de Agrimensura"                  
#> [121] "Universidad de Buenos Aires - Universidad de Buenos Aires"                                                                                    
#> [122] "Universidad Nacional de Entre Ríos - Universidad Nacional de Entre Ríos (IDEER)"                                                              
#> [123] "Universidad Nacional de General Sarmiento - Instituto del Conurbano (IDECONURBANO)"                                                           
#> [124] "Universidad Nacional de La Pampa - Universidad Nacional de La Pampa (IDEUNLPam)"                                                              
#> [125] "Universidad Nacional de la Patagonia San Juan Bosco - Universidad Nacional de la Patagonia San Juan Bosco"                                    
#> [126] "Universidad Nacional de la Patagonia San Juan Bosco - Universidad Nacional de la Patagonia San Juan Bosco - Sede Trelew (IDEUNP)"             
#> [127] "Universidad Nacional de San Juan - Centro de Investigaciones de la Geósfera y la Biósfera (CIGEOBIO) (CONICET-UNSJ)"                          
#> [128] "Universidad Nacional del Centro de la Provincia de Buenos Aires - Universidad Nacional del Centro de la Provincia de Buenos Aires (UNICEN)"   
#> [129] "Universidad Nacional del Sur - Universidad Nacional del Sur (IDESOB)"                                                                         
#> [130] "Universidad Nacional del Comahue - Universidad Nacional del Comahue (IDEUNCO)"                                                                
#> Non-interactive mode: Selecting the first organization.
print(selected_org)
#> # A tibble: 1 × 3
#>   Name                                                           WMS_URL WFS_URL
#>   <chr>                                                          <chr>   <chr>  
#> 1 "Jefatura de Gabinete de Ministros - Comisión Nacional de Act… https:… https:…

if(interactive()){
  # Select an organization interactively
  selected_org <- argentum_select_organization()

  # Select an organization with a search term
  selected_org <- argentum_select_organization("Example")
}